refactor: move loadingscreen.tsx component
