+++ src/components/application/index.ts
@@ -4,5 +4,5 @@ export { injectAppServices } from "./injectAppServices";
 export { useAppServices } from "./useAppServices";
 export { useAppSessionUserData } from "./useAppSessionUserData";
 export { RequireAuth } from "./RequireAuth";
-export { LoadingScreen } from "./LoadingScreen";
+export { LoadingScreen } from "../shared/Loading/LoadingScreen";
 export { NavigateToExternalUrl } from "./NavigateToExternalUrl";
